注解校验完善，修复已知问题

+++ coolGuard-system/src/main/java/cn/wnhyang/coolGuard/system/service/impl/DictServiceImpl.java
@@ -54,7 +54,13 @@ public class DictServiceImpl implements DictService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void delete(Long id) {
-        validateDictExists(id);
+        Dict dict = dictMapper.selectById(id);
+        if (dict == null) {
+            throw exception(DICT_NOT_EXISTS);
+        }
+        if (dict.getStandard()) {
+            throw exception(DICT_STANDARD_NOT_ALLOW_DELETE);
+        }
         dictMapper.deleteById(id);
     }
